+++ crates/mapache/src/common/mod.rs
@@ -5,6 +5,7 @@ pub mod global;
 pub mod hash;
 pub(crate) mod hooks;
 pub mod id;
+pub mod kdf;
 pub mod traits;
 pub mod vars;